dom/plugins/test/mochitest/test_GCrace.html

<head>
  <title>GC race with actors on the parent</title>

  <script type="text/javascript"
	  src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="utils.js"></script>
  <link rel="stylesheet" type="text/css"
	href="/tests/SimpleTest/test.css" />
<body onload="start()">
  <p id="display"></p>

  <script class="testbody" type="application/javascript">
    SimpleTest.waitForExplicitFinish();
    setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_ENABLED);

    function start() {
      if (!SimpleTest.testPluginIsOOP()) {
        ok(true, "Skipping this test when test plugin is not OOP.");
        SimpleTest.finish();
        return;
      }
      else {
        if (navigator.platform.startsWith("Win")) {
          SimpleTest.expectAssertions(0, 350);
        }

        setTimeout(checkGCRace, 1000);
      }
    }

    var nested = false;

    function cb(f) {
      ok(!nested, "Callback shouldn't occur in a nested stack frame");
      try {
        f(35);
        ok(true, "Callback was called, no crash");
      }
      catch (e) {
        ok(false, "Exception calling callback object: " + e);
      }
      SimpleTest.executeSoon(removePlugin);
    }

    function removePlugin() {
      var p = document.getElementById('p');
      p.parentNode.removeChild(p);
      p = null;
      SpecialPowers.Cu.forceGC();
      SimpleTest.finish();
    }

    function checkGCRace() {
      nested = true;

      // The plugin will hand back a function and immediately sleep.
      // We will lose our only reference to the function and force GC, followed
      // by calling us with that function object again. We should be able to
      // call the function and not crash.
      var p = document.getElementById('p');
      var f = p.checkGCRace(cb);
      f = null;  // 'f' should be collected next GC

      nested = false;

      setTimeout(function() {
        SpecialPowers.Cu.forceGC();
      }, 2000);
    }
  </script>

  <embed id="p" type="application/x-test" wmode="window"></embed>
